piwatcher/base/pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "piwatcher_base"
version = "0.1.0"
description = "A robust RTSP video management system for Raspberry Pi."
authors = [{ name = "Bryan Ostdiek", email = "bostdiek@gmail.com" }]
dependencies = ["python-dotenv", "fastapi", "uvicorn", "opencv-python"]
python = ">=3.9"

[tool.poetry]
name = "piwatcher_base"
version = "0.1.0"
description = "A robust RTSP video management system for Raspberry Pi."
authors = ["Bryan Ostdiek <bostdiek@gmail.com>"]


[tool.poetry.dependencies]
python = ">=3.9"

# Optional: Declare test dependencies for the base project if separate tests exist for it
[tool.poetry.group.dev.dependencies]
pytest = "^6.0"
pytest-cov = "^4.0"
coverage = "^7.0"

# Formatting tools
[tool.black]
line-length = 120


[tool.isort]
profile = "black"
line_length = 120
